reworked ipp remap
